AHHHH IRELAND CODING
we now have some basics in the ir transform. we still need to do
some validations, like that there arent duplicate fields in the
keys, indexes or uniques. we should also check that the related
field has an appropriate type, maybe?

also we need to do crud transformations, still. wowzers!